sage/backend/pyproject.toml
[project]
name = "sage"
version = "0.1.0"
description = "Sage - AI Executive Assistant"
authors = [{ name = "Dave Loeffel" }]
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # Vector database
    "qdrant-client>=1.12.0",

    # Cache
    "redis>=5.2.0",

    # AI
    "anthropic>=0.40.0",
    "sentence-transformers>=3.0.0",

    # Google APIs
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.155.0",

    # MCP
    "fastmcp>=2.0.0",

    # Scheduling
    "apscheduler>=3.10.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # HTTP client
    "httpx>=0.28.0",

    # Utilities
    "python-multipart>=0.0.18",
    "python-dotenv>=1.0.1",
    "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sage"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
